Created variables to store operations values
Using an array to operate on the numbers wasn't working, so now the values are stored in number1 and number2, and the operator in operator.
With that, when the handleOperatorsEvents() is called, the program calculated the equation depending on the operator used, and displays the result in the display screen.